[FEAT] 구조 통합 (#125)
