+++ FoodBowl/Data/DTO/Request/GetReviewByStoreRequest.swift
