+++ FoodBowl/Data/DTO/Request/SearchStoresRequest.swift
