+++ FoodBowl/Data/DTO/Request/SignRequestDTO.swift
@@ -1,5 +1,5 @@
 //
-//  SignRequest.swift
+//  SignRequestDTO.swift
 //  FoodBowl
 //
 //  Created by COBY_PRO on 2023/08/10.
