+++ FoodBowl/Data/DTO/Response/CategoriesResponse.swift
