+++ FoodBowl/Data/DTO/Response/FollowMemberResponse.swift
