+++ FoodBowl/Data/DTO/Response/NicknameExistResponse.swift
