+++ FoodBowl/Data/DTO/Response/PlaceResponse.swift
