+++ FoodBowl/Data/DTO/Response/ReviewByStoreResponse.swift
