+++ FoodBowl/Data/DTO/Response/SchoolResponse.swift
