+++ FoodBowl/Data/DTO/Response/SearchStoresResponse.swift
