+++ FoodBowl/Data/DTO/Response/SignDTO.swift
