+++ FoodBowl/Data/DTO/Response/StoreResponse.swift
