--- FoodBowl/Presentation/Common/Base/BaseScrollView.swift
@@ -1,32 +1,0 @@
-//
-//  BaseScrollView.swift
-//  FoodBowl
-//
-//  Created by COBY_PRO on 2022/12/25.
-//
-
-import UIKit
-
-class BaseScrollView<Model>: UIScrollView {
-    var model: Model? {
-        didSet {
-            if let model = model {
-                bind(model)
-            }
-        }
-    }
-
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-
-        configureUI()
-    }
-
-    @available(*, unavailable)
-    required init?(coder _: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    func configureUI() {}
-    func bind(_: Model) {}
-}
