+++ FoodBowl/Presentation/Scene/Map/View/UIComponent/FeedListView.swift
@@ -38,6 +38,7 @@ final class FeedListView: UIView, BaseViewType {
         $0.register(FeedCollectionViewCell.self, forCellWithReuseIdentifier: FeedCollectionViewCell.className)
         $0.backgroundColor = .mainBackgroundColor
     }
+    let emptyView = EmptyView()
     
     // MARK: - property
     
@@ -67,7 +68,7 @@ final class FeedListView: UIView, BaseViewType {
     // MARK: - base func
     
     func setupLayout() {
-        addSubviews(borderLineView, listCollectionView)
+        addSubviews(borderLineView, listCollectionView, emptyView)
 
         borderLineView.snp.makeConstraints {
             $0.top.leading.trailing.equalToSuperview()
@@ -75,6 +76,11 @@ final class FeedListView: UIView, BaseViewType {
         }
 
         listCollectionView.snp.makeConstraints {
+            $0.top.equalTo(borderLineView.snp.bottom)
+            $0.leading.trailing.bottom.equalToSuperview()
+        }
+        
+        emptyView.snp.makeConstraints {
             $0.top.equalTo(borderLineView.snp.bottom)
             $0.leading.trailing.bottom.equalToSuperview()
         }
