+++ FoodBowl/Presentation/Scene/StoreDetail/View/UIComponent/StoreHeaderView.swift
