+++ FoodBowl/Utils/Literal/ImageLiteral.swift
@@ -31,7 +31,6 @@ enum ImageLiteral {
 
     static var camera: UIImage { .load(name: "camera") }
 
-    static var vegan: UIImage { .load(name: "vegan") }
     static var cafe: UIImage { .load(name: "cafe") }
     static var pub: UIImage { .load(name: "pub") }
     static var korean: UIImage { .load(name: "korean") }
